mat mult OpenCL impl running

+++ src/matmut.cl
@@ -1,7 +1,7 @@
 __kernel void matmut(
-  __global double *inputA,
-  __global double *inputB,
-  __global double *outputC,
+  __global float *inputA,
+  __global float *inputB,
+  __global float *outputC,
   int widthA,
   int widthB,
   int heightA,
@@ -10,7 +10,7 @@ __kernel void matmut(
     int row = get_global_id(0);
     int col = get_global_id(1);
 
-    double sum = 0.0;
+    float sum = 0.0;
 
     for(int i = 0; i < widthA; i++) {
       sum += inputA[row * widthA + i] * inputB[i*widthB + col];
